Make arduino wrappers & services more easily-includable
Seperate the arduino wrappers, and services (reset button,
CDC-ACM for Serial class, etc) into separate directories, and arrange so
that users don need to inter-mingle their code with existing zephyr code
to use the services in the application.

To avoid duplication, the create_symlinks.sh script is provided for the
CODK-M Makefile to use during setup. This script will create a symlink to
the arduino & services directory in each example app, so that they
will compile without any effort required from the user.